Further code cleanup
