+++ simple-chat-client/src/lib.rs
@@ -248,3 +248,34 @@ pub async fn client_app(args: Args) {
     let _ = process_server_handle.await;
     let _ = process_terminal.await;
 }
+
+#[cfg(test)]
+mod test {
+    use simple_chat_shared::ClientServerMessages;
+
+    use crate::convert_string_to_command;
+
+    #[tokio::test]
+    async fn test_valid_string_to_command() {
+        let username = "Test";
+        let command_message = "send This is a test message";
+
+        let command_result = convert_string_to_command(username, command_message).unwrap();
+
+        assert_eq!(
+            command_result,
+            ClientServerMessages::Message(
+                username.to_string(),
+                "This is a test message".to_string()
+            )
+        )
+    }
+
+    #[tokio::test]
+    async fn test_invalid_string_to_command() {
+        let username = "Test";
+        let command_message = "Test This is a test message";
+
+        convert_string_to_command(username, command_message).unwrap_err();
+    }
+}
